refactor(auth, consult): 패키지 구조 통일 및 common 제거
- medication과 동일한 패키지 구조(mapper, model) 적용
- 모호한 역할의 common 패키지 삭제 및 파일 재배치
- 프로젝트 전반의 아키텍처 일관성 확보

Closes #21
