+++ features/profile/src/main/java/com/moon/pharm/profile/auth/screen/SignUpScreen.kt
@@ -50,7 +50,7 @@ import com.moon.pharm.component_ui.theme.White
 import com.moon.pharm.domain.model.auth.UserType
 import com.moon.pharm.domain.model.pharmacy.Pharmacy
 import com.moon.pharm.profile.R
-import com.moon.pharm.profile.auth.common.SignUpUiMessage
+import com.moon.pharm.profile.auth.model.SignUpUiMessage
 import com.moon.pharm.profile.auth.mapper.asString
 import com.moon.pharm.profile.auth.model.SignUpStep
 import com.moon.pharm.profile.auth.screen.section.EmailPasswordSection
